fix(pos-login): placed login info next to userid field
