+++ apps/point-of-sale/src/components/Keypad/KeypadComponent.vue
@@ -8,25 +8,9 @@
       </div>
     </div>
   </div>
-
-  <div class="fs-5 text-decoration-underline text-center" @click="() => howToModalVisible = true">How do I login?</div>
-
-  <Dialog v-model:visible="howToModalVisible" header="How do I login?" modal :draggable="false" ref="howToModal"
-    @show="addListenerOnDialogueOverlay(howToModal)">
-    <div class="d-flex flex-column align-items-center">
-      <div>
-        <b style="font-weight: bold!important;">User id:</b> You can use your GEWIS number as user id. <br>
-        <b style="font-weight: bold!important;">Pin:</b> You can (re)set your pin at https://sudosos.gewis.nl/ <br>
-      </div>
-      <img src="@/assets/sudosos-qr.png" style="width: 25rem;" />
-    </div>
-  </Dialog>
 </template>
 
 <script setup lang="ts">
-import { ref, type Ref } from 'vue';
-import { addListenerOnDialogueOverlay } from "@/utils/dialogUtil";
-
 const keypadBackspace = 'B';
 const keypadExternal = 'E';
 const keypadLayout = [
@@ -36,8 +20,6 @@ const keypadLayout = [
   ['E', '0', 'B']
 ];
 
-const howToModalVisible: Ref<boolean> = ref<boolean>(false);
-const howToModal: Ref<null | any> = ref(null);
 
 const emits = defineEmits(['backspace', 'continue', 'input', 'external']);
